Install only do new release when new pulls are available

+++ install.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
-git pull origin master
-rm *.{css,js}_* *.ico.*
-groovy index.groovy
+git remote update
+if [[ `git status -uno | grep "# Your branch is behind"` ]] ; then 
+	echo "We are behind, building a new release..."
+	git pull origin master
+	rm *.{css,js}_* *.ico.* 
+	groovy index.groovy
+        curl -X PURGE http://deigote.com/about-me
+else 
+	echo "Everything is up to date"
+fi
